requirements.txt
chainlit
langchain-google-genai
langchain-ollama
langchain-mcp-adapters
langgraph
python-dotenv
mcp
pydantic>=2.0.0,<3.0.0
requests>=2.31.0
enum34>=1.1.10;python_version<"3.4"